Add viewport
The game world can be larger than what is rendered on screen. This the
concept of a viewport which is the subsection of the game that is
rendered. When the character is more than 100px from the center, the
viewport scrolls so that the character is back in the center.